Safari-609.1.11.1/LayoutTests/svg/animations/animate-path-animation-cC-sS-inverse-expected.txt

SVG 1.1 dynamic animation tests

Test path animation where coordinate modes of start and end differ. You should see PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.animatedPathSegList.numberOfItems is 4
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is -20
PASS path.animatedPathSegList.getItem(0).y is -20
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "c"
PASS path.animatedPathSegList.getItem(1).x is 40
PASS path.animatedPathSegList.getItem(1).y is 40
PASS path.animatedPathSegList.getItem(1).x1 is 0
PASS path.animatedPathSegList.getItem(1).y1 is 40
PASS path.animatedPathSegList.getItem(1).x2 is 0
PASS path.animatedPathSegList.getItem(1).y2 is 40
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "s"
PASS path.animatedPathSegList.getItem(2).x is 0
PASS path.animatedPathSegList.getItem(2).y is -40
PASS path.animatedPathSegList.getItem(2).x2 is 40
PASS path.animatedPathSegList.getItem(2).y2 is 0
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 4
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is -20
PASS path.pathSegList.getItem(0).y is -20
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "c"
PASS path.pathSegList.getItem(1).x is 40
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(1).x1 is 0
PASS path.pathSegList.getItem(1).y1 is 40
PASS path.pathSegList.getItem(1).x2 is 0
PASS path.pathSegList.getItem(1).y2 is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "s"
PASS path.pathSegList.getItem(2).x is 0
PASS path.pathSegList.getItem(2).y is -40
PASS path.pathSegList.getItem(2).x2 is 40
PASS path.pathSegList.getItem(2).y2 is 0
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.animatedPathSegList.numberOfItems is 4
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is -20
PASS path.animatedPathSegList.getItem(0).y is -20
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "c"
PASS path.animatedPathSegList.getItem(1).x is 40
PASS path.animatedPathSegList.getItem(1).y is 40
PASS path.animatedPathSegList.getItem(1).x1 is 10
PASS path.animatedPathSegList.getItem(1).y1 is 30
PASS path.animatedPathSegList.getItem(1).x2 is 10
PASS path.animatedPathSegList.getItem(1).y2 is 30
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "s"
PASS path.animatedPathSegList.getItem(2).x is -10
PASS path.animatedPathSegList.getItem(2).y is -30
PASS path.animatedPathSegList.getItem(2).x2 is 30
PASS path.animatedPathSegList.getItem(2).y2 is 5
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 4
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is -20
PASS path.pathSegList.getItem(0).y is -20
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "c"
PASS path.pathSegList.getItem(1).x is 40
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(1).x1 is 0
PASS path.pathSegList.getItem(1).y1 is 40
PASS path.pathSegList.getItem(1).x2 is 0
PASS path.pathSegList.getItem(1).y2 is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "s"
PASS path.pathSegList.getItem(2).x is 0
PASS path.pathSegList.getItem(2).y is -40
PASS path.pathSegList.getItem(2).x2 is 40
PASS path.pathSegList.getItem(2).y2 is 0
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.animatedPathSegList.numberOfItems is 4
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is -20
PASS path.animatedPathSegList.getItem(0).y is -20
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "C"
PASS path.animatedPathSegList.getItem(1).x is 20
PASS path.animatedPathSegList.getItem(1).y is 20
PASS path.animatedPathSegList.getItem(1).x1 is 10
PASS path.animatedPathSegList.getItem(1).y1 is -10
PASS path.animatedPathSegList.getItem(1).x2 is 10
PASS path.animatedPathSegList.getItem(1).y2 is -10
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "S"
PASS path.animatedPathSegList.getItem(2).x is -10
PASS path.animatedPathSegList.getItem(2).y is 10
PASS path.animatedPathSegList.getItem(2).x2 is 30
PASS path.animatedPathSegList.getItem(2).y2 is 35
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 4
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is -20
PASS path.pathSegList.getItem(0).y is -20
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "c"
PASS path.pathSegList.getItem(1).x is 40
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(1).x1 is 0
PASS path.pathSegList.getItem(1).y1 is 40
PASS path.pathSegList.getItem(1).x2 is 0
PASS path.pathSegList.getItem(1).y2 is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "s"
PASS path.pathSegList.getItem(2).x is 0
PASS path.pathSegList.getItem(2).y is -40
PASS path.pathSegList.getItem(2).x2 is 40
PASS path.pathSegList.getItem(2).y2 is 0
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.animatedPathSegList.numberOfItems is 4
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is -20
PASS path.animatedPathSegList.getItem(0).y is -20
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "C"
PASS path.animatedPathSegList.getItem(1).x is 20
PASS path.animatedPathSegList.getItem(1).y is 20
PASS path.animatedPathSegList.getItem(1).x1 is 20
PASS path.animatedPathSegList.getItem(1).y1 is -20
PASS path.animatedPathSegList.getItem(1).x2 is 20
PASS path.animatedPathSegList.getItem(1).y2 is -20
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "S"
PASS path.animatedPathSegList.getItem(2).x is -20
PASS path.animatedPathSegList.getItem(2).y is 20
PASS path.animatedPathSegList.getItem(2).x2 is 20
PASS path.animatedPathSegList.getItem(2).y2 is 40
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 4
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is -20
PASS path.pathSegList.getItem(0).y is -20
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "c"
PASS path.pathSegList.getItem(1).x is 40
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(1).x1 is 0
PASS path.pathSegList.getItem(1).y1 is 40
PASS path.pathSegList.getItem(1).x2 is 0
PASS path.pathSegList.getItem(1).y2 is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "s"
PASS path.pathSegList.getItem(2).x is 0
PASS path.pathSegList.getItem(2).y is -40
PASS path.pathSegList.getItem(2).x2 is 40
PASS path.pathSegList.getItem(2).y2 is 0
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.animatedPathSegList.numberOfItems is 4
PASS path.animatedPathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.animatedPathSegList.getItem(0).x is -20
PASS path.animatedPathSegList.getItem(0).y is -20
PASS path.animatedPathSegList.getItem(1).pathSegTypeAsLetter is "c"
PASS path.animatedPathSegList.getItem(1).x is 40
PASS path.animatedPathSegList.getItem(1).y is 40
PASS path.animatedPathSegList.getItem(1).x1 is 0
PASS path.animatedPathSegList.getItem(1).y1 is 40
PASS path.animatedPathSegList.getItem(1).x2 is 0
PASS path.animatedPathSegList.getItem(1).y2 is 40
PASS path.animatedPathSegList.getItem(2).pathSegTypeAsLetter is "s"
PASS path.animatedPathSegList.getItem(2).x is 0
PASS path.animatedPathSegList.getItem(2).y is -40
PASS path.animatedPathSegList.getItem(2).x2 is 40
PASS path.animatedPathSegList.getItem(2).y2 is 0
PASS path.animatedPathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS path.pathSegList.numberOfItems is 4
PASS path.pathSegList.getItem(0).pathSegTypeAsLetter is "M"
PASS path.pathSegList.getItem(0).x is -20
PASS path.pathSegList.getItem(0).y is -20
PASS path.pathSegList.getItem(1).pathSegTypeAsLetter is "c"
PASS path.pathSegList.getItem(1).x is 40
PASS path.pathSegList.getItem(1).y is 40
PASS path.pathSegList.getItem(1).x1 is 0
PASS path.pathSegList.getItem(1).y1 is 40
PASS path.pathSegList.getItem(1).x2 is 0
PASS path.pathSegList.getItem(1).y2 is 40
PASS path.pathSegList.getItem(2).pathSegTypeAsLetter is "s"
PASS path.pathSegList.getItem(2).x is 0
PASS path.pathSegList.getItem(2).y is -40
PASS path.pathSegList.getItem(2).x2 is 40
PASS path.pathSegList.getItem(2).y2 is 0
PASS path.pathSegList.getItem(3).pathSegTypeAsLetter is "Z"
PASS successfullyParsed is true

TEST COMPLETE
